move the `get_bbox` to be a property in `LineGeometry` class
