+++ src/dfastbe/io/config.py
@@ -1151,6 +1151,32 @@ class ConfigFile:
         return output_dir
 
 
+def get_bbox(
+    coords: np.ndarray, buffer: float = 0.1
+) -> Tuple[float, float, float, float]:
+    """
+    Derive the bounding box from a line.
+
+    Args:
+        coords (np.ndarray):
+            An N x M array containing x- and y-coordinates as first two M entries
+        buffer : float
+            Buffer fraction surrounding the tight bounding box
+
+    Returns:
+        bbox (Tuple[float, float, float, float]):
+            Tuple bounding box consisting of [min x, min y, max x, max y)
+    """
+    x = coords[:, 0]
+    y = coords[:, 1]
+    x_min = x.min()
+    y_min = y.min()
+    x_max = x.max()
+    y_max = y.max()
+    d = buffer * max(x_max - x_min, y_max - y_min)
+    bbox = (x_min - d, y_min - d, x_max + d, y_max + d)
+
+    return bbox
 
 
 def _move_parameter_location(
